refactor/solve merge conflicts
