+++ src/app.ts
@@ -3,8 +3,8 @@ import cors from "cors";
 import "reflect-metadata";
 import "./di-container";
 import healthRoutes from "./contexts/SystemHealth/presentation/http/routes/healthRoutes";
-import freelancersRoutes from "./contexts/CoreContext/presentation/http/routes/FreelancerRoutes";
 import courseRoutes from "./contexts/LearningContext/presentation/http/routes/CourseRoutes";
+import freelancersRoutes from "./contexts/CoreContext/presentation/http/routes/FreelancersRoutes";
 import { ErrorHandlerMiddleware } from "./contexts/Shared/infrastructure/middlewares/ErrorHandlerMiddleware";
 import userRoutes from "./contexts/CoreContext/presentation/http/routes/UserRoutes";
 import swaggerUi from "swagger-ui-express";
